email module
this script includes the email module, which will actually allow you to send the scam email directly to your inbox (hopefully your spam folder!!). I have made a separate email account specifically for this purpose, scamlibs@gmail.com, however, I am unsure how to publish this code without revealing the email password. So to remedy this, in this specific version you must write your own email (must be a gmail account) & password associated with this account directly into the source code. Also, you must 'enable less secure apps' on this account in google settings. Please note: you will have to enter 2 emails, one directly in the source code (which is where the email will be sent from), and one as the program is running (which is where the email will be sent to). You only need to enter the password associated with the sending email in the source code. When I eventually publish the version associated with scamlibs@gmail.com, all emails will be sent from this address, and you will only need to enter the address you would like the email to be sent to.   Please feel free to email me with any questions.